fix(outlook): make the shipped quoted-history path the tested one
The reduction lived twice: _reduce_message_body, which 58 tests drove and
nothing called, and _exclude_quoted_history, which shipped. They had already
drifted apart in guard order and in how each decided whether a value was
markup, and the copy that ships was the one that got it wrong.

Collapsing both onto plan_body_replacement/apply_body_replacement in a new
outlook_mime module leaves one implementation, and turns four defects into
deletions rather than patches:

- uniqueBody arrives as a raw dict on office365 2.x and as a typed ItemBody
  carrying an enum contentType on 3.x. The connector read only the dict, so on
  3.x every message logged a warning and kept its quoted history. 3.x is the
  current release and the outlook extra declared no bound, so a published
  install already resolved to it. UniqueBody.from_graph reads either shape and
  the extra now pins >=2.6.2,<4.
- A missing or null contentType short-circuited the rendering check and put
  HTML into a text/plain part, which is exactly what that check exists to
  prevent. A rendering is now a BodyRendering or nothing, and nothing is not
  the rendering that was asked for.
- A content value that was not a string raised out of the downloader and
  discarded a download that was already complete on disk. It is a decline now.
- A stylesheet counted as visible text, so a body that says nothing could
  replace a body that says something.

Two more surfaced while moving the code:

- A body whose text carried one of the message's own MIME boundaries forged a
  part. The generator reuses a boundary the message already has without
  scanning the payload for it, and neither 7bit nor quoted-printable escapes a
  leading "--", so the message silently gained a part or ended early with an
  attachment left unreachable. Any container whose delimiter the new text could
  forge now gets a fresh boundary.
- The sweep had no multipart/related carve-out, so a cid-addressed text sibling
  could be removed and leave its reference dangling.

The tests move with the code and drive the shipped path. test_outlook.py is
split three ways with the shared message shapes in outlook_messages.py, so no
file crosses a thousand lines.

Version goes to 1.11.12 rather than 1.11.13: this is not stacked on the CI
fixture-seeding branch, and merging first should not leave a hole.
