+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [1.11.13]
+## [1.11.12]
 
 ### Enhancements
