+++ test/unit/connectors/test_outlook.py
@@ -1,41 +1,20 @@
-import email
-import email.policy
 import hashlib
-import logging
-import re
 from datetime import datetime, timezone
-from email.message import EmailMessage
-from pathlib import Path
 from typing import Optional
 from unittest.mock import MagicMock, Mock, patch
 
 import pytest
 from pydantic import Secret
 
-from unstructured_ingest.data_types.file_data import (
-    FileData,
-    FileDataSourceMetadata,
-    SourceIdentifiers,
-)
 from unstructured_ingest.error import ValueError
 from unstructured_ingest.processes.connectors.outlook import (
-    BODY_PART_PREFERENCE,
     MESSAGE_SELECT_FIELDS,
     MESSAGES_PAGE_SIZE,
-    UNIQUE_BODY_FIELD,
     OutlookAccessConfig,
     OutlookConnectionConfig,
-    OutlookDownloader,
-    OutlookDownloaderConfig,
     OutlookIndexer,
     OutlookIndexerConfig,
-    _has_visible_text,
-    _is_unnamed_text_part,
-    _parse,
-    _prefer_body_rendering,
     _prefer_immutable_ids,
-    _primary_body_part,
-    _reduce_message_body,
 )
 
 
@@ -447,1014 +426,3 @@ class TestPreferImmutableIdsHeader:
         )
         client.pending_request().beforeExecute.notify(continuation_request)
         assert continuation_request.headers["Prefer"] == 'IdType="ImmutableId"'
-
-
-EML_DIR = Path(__file__).resolve().parents[3] / "example-docs" / "eml"
-
-# Committed fixtures chosen for MIME structure rather than content.
-HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT = EML_DIR / "email-with-image.eml"
-HTML_AND_PLAIN_WITH_TEXT_ATTACHMENT = EML_DIR / "fake-email-attachment.eml"
-# Carries two attachments and a long non-ASCII Authentication-Results header,
-# which is the only shape that catches header refolding on re-serialization.
-LONG_NON_ASCII_HEADERS = EML_DIR / "email-no-utf8-2014-03-17.111517.eml"
-# multipart/related: the inline image is a sibling of the body part.
-INLINE_IMAGE_RELATED = EML_DIR / "fake-email-image-embedded.eml"
-
-STRUCTURED_FIXTURES = [
-    HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT,
-    HTML_AND_PLAIN_WITH_TEXT_ATTACHMENT,
-    LONG_NON_ASCII_HEADERS,
-    INLINE_IMAGE_RELATED,
-]
-
-UNIQUE_HTML = "<html><body><p>Only the newest sentence.</p></body></html>"
-
-PREFER_BODY_RENDERING = "unstructured_ingest.processes.connectors.outlook._prefer_body_rendering"
-
-# Must survive inside an attached message rather than being mistaken for a
-# superseded rendering of the outer body.
-NESTED_SENTINEL = "THE ATTACHED MESSAGE BODY MUST SURVIVE"
-
-SIGNED_MESSAGE = (
-    b'Content-Type: multipart/signed; protocol="application/pkcs7-signature"; '
-    b'micalg=sha-256; boundary="s1"\r\nSubject: signed mail\r\n\r\n'
-    b"--s1\r\nContent-Type: text/plain\r\n\r\nThe signed body text.\r\n"
-    b"--s1\r\nContent-Type: application/pkcs7-signature\r\n\r\nc2lnbmF0dXJl\r\n--s1--\r\n"
-)
-
-# A single-part message: the body part is the message itself, so the surgery
-# rewrites top-level headers rather than a child part's.
-SINGLE_PART_PLAIN = (
-    b"Subject: quarterly update\r\nFrom: sender@example.com\r\nTo: rec@example.com\r\n"
-    b"Content-Type: text/plain; charset=us-ascii\r\n\r\nThe original body text.\r\n"
-)
-
-ATTACHMENT_ONLY = (
-    b'Content-Type: multipart/mixed; boundary="b1"\r\nSubject: attachment only\r\n\r\n'
-    b"--b1\r\nContent-Type: application/pdf\r\n"
-    b'Content-Disposition: attachment; filename="a.pdf"\r\n\r\nnot-a-pdf\r\n--b1--\r\n'
-)
-
-# An empty body plus an empty uniqueBody: the shape Graph returns for a message
-# that carries only attachments in its body slot.
-EMPTY_BODY = (
-    b"Subject: nothing to say\r\nFrom: sender@example.com\r\n"
-    b"Content-Type: text/plain; charset=us-ascii\r\n\r\n\r\n"
-)
-
-
-def _leaves(message):
-    return [part for part in message.walk() if not part.is_multipart()]
-
-
-def _non_body_facts(message) -> list[tuple]:
-    """Identity of every part that is not a body rendering."""
-    return [
-        (
-            part.get_content_type(),
-            part.get_filename(),
-            part.get("Content-Transfer-Encoding"),
-            part.get_payload(decode=True),
-        )
-        for part in _leaves(message)
-        if not _is_unnamed_text_part(part)
-    ]
-
-
-def _header_facts(message) -> list[tuple[str, str]]:
-    """Headers a body replacement has no business changing."""
-    structural = {"content-type", "content-transfer-encoding", "mime-version"}
-    return sorted(
-        (name.lower(), str(value))
-        for name, value in message.items()
-        if name.lower() not in structural
-    )
-
-
-class TestPrimaryBodyPart:
-    """The part reported must be the one a partitioner will read.
-
-    Its subtype decides which rendering Graph is asked for, so a wrong answer
-    here puts the wrong markup into the message.
-    """
-
-    @pytest.mark.parametrize("fixture", STRUCTURED_FIXTURES, ids=lambda p: p.name)
-    def test_prefers_html_when_the_message_carries_both(self, fixture: Path):
-        part = _primary_body_part(fixture.read_bytes())
-        assert part is not None
-        assert part.get_content_subtype() == "html"
-
-    def test_reports_plain_when_there_is_no_html_rendering(self):
-        part = _primary_body_part(SINGLE_PART_PLAIN)
-        assert part is not None
-        assert part.get_content_subtype() == "plain"
-
-    def test_reports_none_when_there_is_no_body_part(self):
-        assert _primary_body_part(ATTACHMENT_ONLY) is None
-
-    def test_excludes_a_named_inline_html_part(self):
-        raw = (
-            b'Content-Type: multipart/mixed; boundary="mix"\r\n\r\n'
-            b"--mix\r\nContent-Type: text/plain\r\n\r\nThe actual message body.\r\n"
-            b"--mix\r\nContent-Type: text/html\r\n"
-            b'Content-Disposition: inline; filename="report.html"\r\n\r\n'
-            b"<p>Attached report contents.</p>\r\n--mix--\r\n"
-        )
-        stdlib = email.message_from_bytes(raw, policy=email.policy.default)
-        assert stdlib.get_body(preferencelist=BODY_PART_PREFERENCE).get_filename() == (
-            "report.html"
-        ), "premise: the standard library selects the named inline part as the body"
-
-        body = _primary_body_part(raw)
-
-        assert body is not None
-        assert body.get_content_type() == "text/plain"
-        assert "actual message body" in body.get_content()
-
-
-class TestReduceMessageBody:
-    """The surgery must change the body and nothing else.
-
-    Losing an attachment here would be worse than the duplication this
-    feature exists to remove, so every invariant gets its own assertion.
-    """
-
-    @pytest.mark.parametrize("fixture", STRUCTURED_FIXTURES, ids=lambda p: p.name)
-    def test_non_body_parts_are_untouched(self, fixture: Path):
-        raw = fixture.read_bytes()
-        before = _non_body_facts(_parse(raw))
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        assert reduced is not None
-        assert _non_body_facts(_parse(reduced)) == before
-
-    def test_a_single_part_message_keeps_its_identifying_headers(self):
-        """The body part is the message itself here, so the surgery rewrites the
-        top-level content type and transfer encoding by design. Everything that
-        identifies the message still has to survive."""
-        reduced = _reduce_message_body(SINGLE_PART_PLAIN, "Only the newest sentence.")
-
-        assert reduced is not None
-        rebuilt, original = _parse(reduced), _parse(SINGLE_PART_PLAIN)
-        for header in ("Subject", "From", "To"):
-            assert str(rebuilt[header]) == str(original[header])
-        assert "Only the newest sentence." in rebuilt.get_content()
-        assert "The original body text." not in rebuilt.get_content()
-
-    @pytest.mark.parametrize("fixture", STRUCTURED_FIXTURES, ids=lambda p: p.name)
-    def test_headers_are_untouched(self, fixture: Path):
-        raw = fixture.read_bytes()
-        before = _header_facts(_parse(raw))
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        assert _header_facts(_parse(reduced)) == before
-
-    def test_long_non_ascii_headers_are_not_refolded(self):
-        """The default serialization policy rewrites long source headers.
-
-        Pinned separately from the parametrized case because a fixture with
-        only short headers cannot catch it: the failure is whitespace inserted
-        inside a header the replacement never touched.
-        """
-        raw = LONG_NON_ASCII_HEADERS.read_bytes()
-        original = _parse(raw)
-        long_headers = [name for name, value in original.items() if len(f"{name}: {value}") > 200]
-        assert long_headers, "fixture no longer carries a long header to protect"
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        rebuilt = _parse(reduced)
-        for name in long_headers:
-            assert str(rebuilt[name]) == str(original[name])
-
-    @pytest.mark.parametrize("fixture", STRUCTURED_FIXTURES, ids=lambda p: p.name)
-    def test_the_body_part_carries_the_supplied_text(self, fixture: Path):
-        reduced = _reduce_message_body(fixture.read_bytes(), UNIQUE_HTML)
-
-        body = _parse(reduced).get_body(preferencelist=BODY_PART_PREFERENCE)
-        assert "Only the newest sentence." in body.get_content()
-
-    @pytest.mark.parametrize("fixture", STRUCTURED_FIXTURES, ids=lambda p: p.name)
-    def test_no_other_body_rendering_survives(self, fixture: Path):
-        """A stale plain-text rendering would still hold the quoted history.
-
-        The partitioner accepts a setting that flips its preference to plain
-        text, which would silently restore that history and make the whole
-        feature a no-op, so the other renderings are removed rather than left.
-        """
-        raw = fixture.read_bytes()
-        assert len([p for p in _leaves(_parse(raw)) if _is_unnamed_text_part(p)]) > 1
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        assert len([p for p in _leaves(_parse(reduced)) if _is_unnamed_text_part(p)]) == 1
-
-    @pytest.mark.parametrize("fixture", STRUCTURED_FIXTURES, ids=lambda p: p.name)
-    def test_the_original_body_text_is_gone(self, fixture: Path):
-        raw = fixture.read_bytes()
-        original_body = _parse(raw).get_body(preferencelist=BODY_PART_PREFERENCE)
-        # Markup is stripped rather than used to skip lines: skipping any line
-        # containing a tag leaves nothing to assert on an HTML body.
-        words = [
-            word
-            for word in re.sub(r"<[^>]+>", " ", original_body.get_content()).split()
-            if len(word) > 8 and word.isalpha()
-        ]
-        assert words, "fixture body has no distinctive words to check"
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        rebuilt_text = _parse(reduced).get_body(preferencelist=BODY_PART_PREFERENCE).get_content()
-        for word in words:
-            assert word not in rebuilt_text
-
-    def test_a_named_inline_html_part_is_not_replaced_or_removed(self):
-        raw = (
-            b'Content-Type: multipart/mixed; boundary="mix"\r\n\r\n'
-            b"--mix\r\nContent-Type: text/plain\r\n\r\nThe actual message body.\r\n"
-            b"--mix\r\nContent-Type: text/html\r\n"
-            b'Content-Disposition: inline; filename="report.html"\r\n\r\n'
-            b"<p>Attached report contents.</p>\r\n--mix--\r\n"
-        )
-
-        reduced = _reduce_message_body(raw, "Only the newest sentence.")
-
-        assert reduced is not None
-        rebuilt = _parse(reduced)
-        named = next(part for part in rebuilt.walk() if part.get_filename() == "report.html")
-        assert named.get_content_disposition() == "inline"
-        assert "Attached report contents." in named.get_content()
-        body = _primary_body_part(reduced)
-        assert body is not None
-        assert "Only the newest sentence." in body.get_content()
-
-    def test_preserves_the_content_id_of_a_related_root(self):
-        raw = (
-            b'Content-Type: multipart/related; boundary="rel"; start="<root>"\r\n\r\n'
-            b"--rel\r\nContent-Type: image/png\r\nContent-ID: <image>\r\n\r\n"
-            b"cG5n\r\n"
-            b"--rel\r\nContent-Type: text/html\r\nContent-ID: <root>\r\n\r\n"
-            b"<p>The full body and quoted history.</p>\r\n--rel--\r\n"
-        )
-        original_body = _parse(raw).get_body(preferencelist=BODY_PART_PREFERENCE)
-        assert original_body is not None
-        assert original_body["Content-ID"] == "<root>"
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        assert reduced is not None
-        rebuilt = _parse(reduced)
-        body = rebuilt.get_body(preferencelist=BODY_PART_PREFERENCE)
-        assert body is not None
-        assert body["Content-ID"] == "<root>"
-        assert "Only the newest sentence." in body.get_content()
-
-
-class TestReduceMessageBodyLeavesNestedMessagesAlone:
-    """An attached message has a body of its own, and it must survive.
-
-    Forwarding mail as an attachment is ordinary Outlook behaviour, and the
-    attached message's body carries no attachment disposition and no filename,
-    so it looks exactly like a superseded rendering of the outer body. Removing
-    body parts anywhere but the outer body's own container guts the attachment.
-    """
-
-    def _message_with_an_attached_message(self) -> bytes:
-        outer = EmailMessage()
-        outer["Subject"] = "please see the attached mail"
-        outer["From"] = "sender@example.com"
-        outer.set_content("outer plain body")
-        outer.add_alternative("<p>outer html body</p>", subtype="html")
-
-        attached = EmailMessage()
-        attached["Subject"] = "the forwarded message"
-        attached["From"] = "original@example.com"
-        attached.set_content(NESTED_SENTINEL)
-
-        # add_attachment of a message object yields a message/rfc822 part,
-        # which is the shape Outlook produces for mail forwarded as a file.
-        outer.add_attachment(attached)
-        assert any(
-            part.get_content_type() == "message/rfc822" for part in _parse(outer.as_bytes()).walk()
-        )
-        return outer.as_bytes()
-
-    def test_the_attached_message_body_survives(self):
-        raw = self._message_with_an_attached_message()
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        assert reduced is not None
-        assert NESTED_SENTINEL in reduced.decode("utf-8", "replace")
-
-    def test_only_the_outer_bodys_own_rendering_is_removed(self):
-        raw = self._message_with_an_attached_message()
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        before = len(_leaves(_parse(raw)))
-        after = len(_leaves(_parse(reduced)))
-        assert after == before - 1
-
-    def test_the_outer_body_is_still_replaced(self):
-        raw = self._message_with_an_attached_message()
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        body = _parse(reduced).get_body(preferencelist=BODY_PART_PREFERENCE)
-        assert "Only the newest sentence." in body.get_content()
-        assert "outer plain body" not in reduced.decode("utf-8", "replace")
-
-
-class TestReduceMessageBodySweepsNestedContainers:
-    """The superseded rendering can sit in a different container.
-
-    A message whose HTML body lives in a related group alongside its images,
-    with the plain-text rendering in a sibling alternative group, is ordinary
-    mail. Sweeping only the body part's own container would leave that plain
-    rendering, quoted history and all, for a plain-preferring partitioner to
-    read, and the setting would silently do nothing.
-    """
-
-    def _nested(self) -> bytes:
-        raw = (
-            b'Content-Type: multipart/mixed; boundary="mix"\r\n'
-            b"Subject: nested renderings\r\n\r\n"
-            b'--mix\r\nContent-Type: multipart/alternative; boundary="alt"\r\n\r\n'
-            b"--alt\r\nContent-Type: text/plain\r\n\r\n"
-            b"PLAINRENDERING with the whole quoted history.\r\n"
-            b"--alt--\r\n"
-            b'--mix\r\nContent-Type: multipart/related; boundary="rel"\r\n\r\n'
-            b"--rel\r\nContent-Type: text/html\r\n\r\n"
-            b"<p>HTMLRENDERING with the whole quoted history.</p>\r\n"
-            b"--rel\r\nContent-Type: image/png\r\nContent-ID: <img1>\r\n\r\n"
-            b"cG5nYnl0ZXM=\r\n--rel--\r\n--mix--\r\n"
-        )
-        assert _parse(raw).get_body(preferencelist=BODY_PART_PREFERENCE).get_content_type() == (
-            "text/html"
-        )
-        return raw
-
-    def test_the_plain_rendering_in_another_container_is_removed(self):
-        reduced = _reduce_message_body(self._nested(), UNIQUE_HTML)
-
-        assert reduced is not None
-        assert b"PLAINRENDERING" not in reduced
-
-    def test_the_inline_image_survives(self):
-        reduced = _reduce_message_body(self._nested(), UNIQUE_HTML)
-
-        types = [part.get_content_type() for part in _leaves(_parse(reduced))]
-        assert "image/png" in types
-
-    def test_no_empty_container_is_left_behind(self):
-        """The alternative group holds nothing once its only rendering goes."""
-        reduced = _reduce_message_body(self._nested(), UNIQUE_HTML)
-
-        containers = [
-            part
-            for part in _parse(reduced).walk()
-            if part.is_multipart() and not part.get_payload()
-        ]
-        assert containers == []
-
-
-class TestReduceMessageBodyKeepsInlineImages:
-    """An image pasted into a reply is carried as a related part, not as data.
-
-    Outlook references it from the body as cid:<id>. The reference lives inside
-    the body text being replaced, and the image itself is a sibling part, so
-    both have to come through: the part because it is not a body rendering, and
-    the reference because Graph keeps it in the value it returns.
-    """
-
-    def _body_with_a_reference(self) -> str:
-        return '<html><body><p>Only the newest sentence.</p><img src="cid:img1"></body></html>'
-
-    def test_the_referenced_image_part_survives(self):
-        raw = INLINE_IMAGE_RELATED.read_bytes()
-
-        reduced = _reduce_message_body(raw, self._body_with_a_reference())
-
-        assert reduced is not None
-        types = [part.get_content_type() for part in _leaves(_parse(reduced))]
-        assert "image/png" in types
-
-    def test_the_reference_itself_survives(self):
-        raw = INLINE_IMAGE_RELATED.read_bytes()
-
-        reduced = _reduce_message_body(raw, self._body_with_a_reference())
-
-        body = _parse(reduced).get_body(preferencelist=BODY_PART_PREFERENCE)
-        assert "cid:img1" in body.get_content()
-
-
-class TestReduceMessageBodyLeavesAttachedContainersAlone:
-    """A container carried as an attachment keeps its own text parts.
-
-    An attached message is recognised by its content type, but a group saved as
-    a file, for instance a related bundle, is only recognisable by its
-    disposition or filename. Descending into one deletes the attachment's text.
-    """
-
-    def _message_with_an_attached_container(self) -> bytes:
-        return (
-            b'Content-Type: multipart/mixed; boundary="mix"\r\n'
-            b"Subject: attached container\r\n\r\n"
-            b'--mix\r\nContent-Type: multipart/alternative; boundary="alt"\r\n\r\n'
-            b"--alt\r\nContent-Type: text/plain\r\n\r\nOuter plain rendering.\r\n"
-            b"--alt\r\nContent-Type: text/html\r\n\r\n"
-            b"<p>Outer html rendering.</p>\r\n--alt--\r\n"
-            b'--mix\r\nContent-Type: multipart/related; boundary="rel"\r\n'
-            b'Content-Disposition: attachment; filename="bundle.mhtml"\r\n\r\n'
-            b"--rel\r\nContent-Type: text/html\r\n\r\n"
-            b"<p>ATTACHEDBUNDLETEXT</p>\r\n"
-            b"--rel\r\nContent-Type: image/png\r\nContent-ID: <b1>\r\n\r\n"
-            b"cG5n\r\n--rel--\r\n--mix--\r\n"
-        )
-
-    def test_the_attached_containers_text_survives(self):
-        raw = self._message_with_an_attached_container()
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        assert reduced is not None
-        assert b"ATTACHEDBUNDLETEXT" in reduced
-
-    def test_the_outer_plain_rendering_is_still_removed(self):
-        raw = self._message_with_an_attached_container()
-
-        reduced = _reduce_message_body(raw, UNIQUE_HTML)
-
-        assert b"Outer plain rendering." not in reduced
-        body = _parse(reduced).get_body(preferencelist=BODY_PART_PREFERENCE)
-        assert "Only the newest sentence." in body.get_content()
-
-
-class TestReduceMessageBodyLineEndings:
-    def test_the_rebuilt_message_uses_crlf(self):
-        """Graph delivers CRLF, and the default policy would flatten it."""
-        reduced = _reduce_message_body(
-            HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes(), UNIQUE_HTML
-        )
-
-        assert reduced is not None
-        assert b"\r\n" in reduced
-        assert reduced.replace(b"\r\n", b"").count(b"\n") == 0
-
-
-class TestReduceMessageBodyDeclines:
-    """Declining must be indistinguishable from the feature being off.
-
-    Graph returns an empty uniqueBody for a legitimately empty body, so the
-    condition cannot simply be "the value is empty".
-    """
-
-    @pytest.mark.parametrize(
-        "unique_content",
-        [None, "", "   \r\n  ", "<html><body><div></div></body></html>"],
-        ids=["missing", "empty", "whitespace", "markup-without-text"],
-    )
-    def test_declines_when_the_value_carries_no_text(self, unique_content):
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-
-        assert _reduce_message_body(raw, unique_content) is None
-
-    def test_declines_when_there_is_no_body_part_to_replace(self):
-        assert _reduce_message_body(ATTACHMENT_ONLY, UNIQUE_HTML) is None
-
-    def test_declines_for_an_entity_only_value(self):
-        """Non-breaking spaces are not text, and would blank a real body.
-
-        The fixture has an HTML body, so the value is read as markup. Against a
-        plain-text body the same string is literal text and is written, which
-        TestCarriesText covers.
-        """
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-
-        assert _reduce_message_body(raw, "<p>&nbsp;&nbsp;</p>") is None
-
-    def test_an_empty_body_and_an_empty_value_leave_the_message_alone(self):
-        assert _reduce_message_body(EMPTY_BODY, "") is None
-
-
-class TestPreferBodyRendering:
-    """Graph names its renderings "text" and "html"; MIME says "plain".
-
-    Passing a MIME subtype through unchanged makes Graph ignore the preference
-    and answer with HTML, which then lands inside a plain-text part.
-    """
-
-    def _request(self):
-        try:
-            from office365.runtime.http.request_options import RequestOptions
-        except ImportError:
-            pytest.skip("office365-rest-python-client not installed")
-        return RequestOptions("https://graph.microsoft.com/v1.0/users/alice/messages/m1")
-
-    def test_translates_the_mime_subtype_for_a_plain_body(self):
-        request = self._request()
-
-        _prefer_body_rendering("plain")(request)
-
-        assert request.headers["Prefer"] == 'outlook.body-content-type="text"'
-
-    def test_passes_html_through(self):
-        request = self._request()
-
-        _prefer_body_rendering("html")(request)
-
-        assert request.headers["Prefer"] == 'outlook.body-content-type="html"'
-
-    def test_every_subtype_the_body_lookup_can_return_is_translatable(self):
-        """The lookup can only ever yield these two, so neither may raise."""
-        for mime_subtype in BODY_PART_PREFERENCE:
-            _prefer_body_rendering(mime_subtype)(self._request())
-
-    def test_composes_with_a_preference_already_set(self):
-        """The immutable-id hook sets Prefer on the same client, so this appends."""
-        request = self._request()
-
-        _prefer_immutable_ids(request)
-        _prefer_body_rendering("html")(request)
-
-        assert request.headers["Prefer"] == 'IdType="ImmutableId", outlook.body-content-type="html"'
-
-
-class TestUniqueBodyRawPropertyLookup:
-    """Pins uniqueBody's shape on the pinned client, like changeKey above.
-
-    The connector reads the raw property because 2.6.2 declares no typed
-    accessor. If an upgrade adds one, this test says so rather than the
-    connector silently reading None and keeping every full body.
-    """
-
-    def _real_message(self):
-        try:
-            from office365.graph_client import GraphClient
-            from office365.outlook.mail.messages.message import Message
-            from office365.runtime.paths.resource_path import ResourcePath
-        except ImportError:
-            pytest.skip("office365-rest-python-client not installed")
-        client = GraphClient(lambda: {"access_token": "x", "token_type": "Bearer"})
-        return Message(client, ResourcePath("messages/abc"))
-
-    def test_there_is_no_typed_accessor(self):
-        assert not hasattr(self._real_message(), "unique_body")
-
-    def test_the_raw_property_holds_the_graph_payload(self):
-        message = self._real_message()
-        payload = {"contentType": "html", "content": "<p>new</p>"}
-
-        message.set_property(UNIQUE_BODY_FIELD, payload)
-
-        assert message.get_property(UNIQUE_BODY_FIELD) == payload
-
-
-class TestOutlookDownloaderConfigDefault:
-    def test_quoted_history_exclusion_is_off_by_default(self):
-        assert OutlookDownloaderConfig().exclude_quoted_history is False
-
-
-class TestCarriesText:
-    """Angle brackets are markup in an HTML body and characters in a text one."""
-
-    @pytest.mark.parametrize(
-        "value",
-        [None, "", "   ", "<div></div>", "<p>&nbsp;</p>", "<p>&#160;&#160;</p>"],
-        ids=["none", "empty", "spaces", "tags", "nbsp-entity", "numeric-entity"],
-    )
-    def test_markup_without_words_is_not_text(self, value):
-        assert _has_visible_text(value, is_markup=True) is False
-
-    @pytest.mark.parametrize(
-        "value",
-        ["hello", "<p>hello</p>", "<p>&amp;</p>"],
-        ids=["bare", "wrapped", "escaped-ampersand"],
-    )
-    def test_markup_with_words_is_text(self, value):
-        assert _has_visible_text(value, is_markup=True) is True
-
-    @pytest.mark.parametrize(
-        "value",
-        ["<no comment>", "<see attached>", "a < b and c > d"],
-        ids=["bracketed-note", "bracketed-pointer", "comparison"],
-    )
-    def test_plain_text_in_angle_brackets_is_still_text(self, value):
-        assert _has_visible_text(value, is_markup=False) is True
-
-    @pytest.mark.parametrize(
-        "value",
-        ["<no comment>", "<see attached>"],
-        ids=["bracketed-note", "bracketed-pointer"],
-    )
-    def test_the_same_values_read_as_empty_markup(self, value):
-        """This is the miss the flag exists to prevent. Read as markup these
-        strip to nothing, so a plain-text message would be left unreduced."""
-        assert _has_visible_text(value, is_markup=True) is False
-
-    @pytest.mark.parametrize("value", [None, "", "  \r\n "], ids=["none", "empty", "whitespace"])
-    def test_plain_text_still_has_to_hold_something(self, value):
-        assert _has_visible_text(value, is_markup=False) is False
-
-
-class TestDownloaderQuotedHistoryRequest:
-    """The extra Graph request happens only when the setting is on."""
-
-    def _downloader(self, exclude_quoted_history: bool):
-        connection_config = OutlookConnectionConfig(
-            access_config=Secret(OutlookAccessConfig(oauth_token="ey.access.token")),
-        )
-        return OutlookDownloader(
-            connection_config=connection_config,
-            download_config=OutlookDownloaderConfig(exclude_quoted_history=exclude_quoted_history),
-        )
-
-    def _file_data(self):
-        return FileData(
-            identifier="msg-1",
-            connector_type="outlook",
-            source_identifiers=SourceIdentifiers(filename="msg-1.eml", fullpath="msg-1.eml"),
-            metadata=FileDataSourceMetadata(
-                record_locator={"message_id": "msg-1", "user_email": "alice@example.com"}
-            ),
-        )
-
-    def _client_writing(self, raw: bytes, unique_content: Optional[str]):
-        """A client whose download writes `raw` and whose select carries uniqueBody.
-
-        The rendering it answers with matches the message's own body part, the
-        way an honest service would, so a test that wants a mismatch has to say
-        so explicitly. Returns the client and the message it hands out, so
-        assertions name the message directly rather than re-walking the mock.
-        """
-        message = MagicMock()
-
-        def _download(file_obj):
-            file_obj.write(raw)
-            return MagicMock()
-
-        body = _primary_body_part(raw)
-        rendering = "text" if body is not None and body.get_content_subtype() == "plain" else "html"
-
-        message.download.side_effect = _download
-        message.select.return_value = message
-        # The pinned office365 client has no typed accessor for uniqueBody, so
-        # the connector reads the raw property dict Graph sent.
-        message.get_property.return_value = (
-            {"contentType": rendering, "content": unique_content}
-            if unique_content is not None
-            else None
-        )
-
-        client = MagicMock()
-        client.users.__getitem__.return_value.messages.__getitem__.return_value = message
-        return client, message
-
-    @pytest.mark.parametrize(
-        "raw",
-        [
-            HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes(),
-            HTML_AND_PLAIN_WITH_TEXT_ATTACHMENT.read_bytes(),
-            LONG_NON_ASCII_HEADERS.read_bytes(),
-            INLINE_IMAGE_RELATED.read_bytes(),
-            SINGLE_PART_PLAIN,
-            SIGNED_MESSAGE,
-            EMPTY_BODY,
-            ATTACHMENT_ONLY,
-        ],
-        ids=[
-            "html-plain-image-attachment",
-            "html-plain-text-attachment",
-            "long-non-ascii-headers",
-            "inline-image-related",
-            "single-part-plain",
-            "signed",
-            "empty-body",
-            "attachment-only",
-        ],
-    )
-    def test_the_setting_off_changes_nothing_at_all(self, tmp_path: Path, raw: bytes):
-        """The default must be indistinguishable from the connector before this.
-
-        Byte-identical output and no extra request, on every message shape the
-        suite knows about, is the whole no-breaking-change claim.
-        """
-        downloader = self._downloader(exclude_quoted_history=False)
-        client, message = self._client_writing(raw, UNIQUE_HTML)
-        download_path = tmp_path / "msg-1.eml"
-
-        with patch.object(OutlookConnectionConfig, "get_client", return_value=client):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert download_path.read_bytes() == raw
-        message.select.assert_not_called()
-        client.execute_query.assert_not_called()
-        assert [path.name for path in tmp_path.iterdir()] == ["msg-1.eml"]
-
-    def test_one_selected_request_when_the_setting_is_on(self, tmp_path: Path):
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, message = self._client_writing(raw, UNIQUE_HTML)
-        download_path = tmp_path / "msg-1.eml"
-
-        with (
-            patch.object(OutlookConnectionConfig, "get_client", return_value=client),
-            patch(PREFER_BODY_RENDERING) as prefer,
-        ):
-            downloader._download_message(self._file_data(), download_path)
-
-        # An HTML body part must be matched by an HTML rendering request.
-        prefer.assert_called_once_with("html")
-
-        message.select.assert_called_once_with(["id", UNIQUE_BODY_FIELD])
-        client.execute_query.assert_called_once()
-
-    def test_the_written_file_carries_only_the_unique_text(self, tmp_path: Path):
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, _ = self._client_writing(raw, UNIQUE_HTML)
-        download_path = tmp_path / "msg-1.eml"
-
-        with patch.object(OutlookConnectionConfig, "get_client", return_value=client):
-            downloader._download_message(self._file_data(), download_path)
-
-        written = download_path.read_bytes()
-        assert written != raw
-        body = _parse(written).get_body(preferencelist=BODY_PART_PREFERENCE)
-        assert "Only the newest sentence." in body.get_content()
-        assert _non_body_facts(_parse(written)) == _non_body_facts(_parse(raw))
-
-    def test_the_full_body_is_kept_when_graph_returns_nothing(self, tmp_path: Path):
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, _ = self._client_writing(raw, None)
-        download_path = tmp_path / "msg-1.eml"
-
-        with patch.object(OutlookConnectionConfig, "get_client", return_value=client):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert download_path.read_bytes() == raw
-
-    def test_one_client_serves_both_requests(self, tmp_path: Path):
-        """A second client would mean a second token acquisition per message."""
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, _ = self._client_writing(raw, UNIQUE_HTML)
-        download_path = tmp_path / "msg-1.eml"
-
-        with patch.object(OutlookConnectionConfig, "get_client", return_value=client) as get_client:
-            downloader._download_message(self._file_data(), download_path)
-
-        get_client.assert_called_once()
-
-    def test_a_lost_reduction_is_warned_about(self, tmp_path: Path, caplog):
-        """A record keeping history it was meant to lose has to be visible."""
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, _ = self._client_writing(raw, None)
-        download_path = tmp_path / "msg-1.eml"
-
-        with (
-            caplog.at_level(logging.WARNING),
-            patch.object(OutlookConnectionConfig, "get_client", return_value=client),
-        ):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert any(record.levelno == logging.WARNING for record in caplog.records)
-
-    def test_an_empty_body_is_not_warned_about(self, tmp_path: Path, caplog):
-        """An attachment-only message has nothing to reduce, so a warning here
-        would make a healthy mailbox look broken."""
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, _ = self._client_writing(EMPTY_BODY, "")
-        download_path = tmp_path / "msg-1.eml"
-
-        with (
-            caplog.at_level(logging.INFO),
-            patch.object(OutlookConnectionConfig, "get_client", return_value=client),
-        ):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert download_path.read_bytes() == EMPTY_BODY
-        assert not [record for record in caplog.records if record.levelno >= logging.WARNING]
-
-    def test_a_plain_only_message_asks_for_the_plain_rendering(self, tmp_path: Path):
-        """A plain body must not be answered with HTML.
-
-        The MIME subtype here is "plain", which Graph does not accept; the
-        translation to its own "text" token happens inside the helper this
-        asserts on, and is covered by TestPreferBodyRendering.
-        """
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, _ = self._client_writing(SINGLE_PART_PLAIN, "Only the newest sentence.")
-        download_path = tmp_path / "msg-1.eml"
-
-        with (
-            patch.object(OutlookConnectionConfig, "get_client", return_value=client),
-            patch(PREFER_BODY_RENDERING) as prefer,
-        ):
-            downloader._download_message(self._file_data(), download_path)
-
-        prefer.assert_called_once_with("plain")
-
-        written = _parse(download_path.read_bytes())
-        assert written.get_content_type() == "text/plain"
-        assert "<" not in written.get_content()
-
-    def test_a_failed_extra_request_does_not_fail_the_record(self, tmp_path: Path, caplog):
-        """The full message is already on disk, so a throttled or failed extra
-        request must leave a usable record rather than raise."""
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, message = self._client_writing(raw, UNIQUE_HTML)
-        client.execute_query.side_effect = RuntimeError("429 too many requests")
-        download_path = tmp_path / "msg-1.eml"
-
-        with (
-            caplog.at_level(logging.WARNING),
-            patch.object(OutlookConnectionConfig, "get_client", return_value=client),
-        ):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert download_path.read_bytes() == raw
-        assert any(record.levelno == logging.WARNING for record in caplog.records)
-
-    def test_a_rendering_graph_did_not_honour_is_refused(self, tmp_path: Path, caplog):
-        """Graph may ignore an unsupported preference and answer in its own
-        default. Writing HTML into a plain-text part would have the partitioner
-        extract literal markup, so the full body is kept instead."""
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, message = self._client_writing(SINGLE_PART_PLAIN, "<p>html we did not ask for</p>")
-        message.get_property.return_value = {
-            "contentType": "html",
-            "content": "<p>html we did not ask for</p>",
-        }
-        download_path = tmp_path / "msg-1.eml"
-
-        with (
-            caplog.at_level(logging.WARNING),
-            patch.object(OutlookConnectionConfig, "get_client", return_value=client),
-        ):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert download_path.read_bytes() == SINGLE_PART_PLAIN
-        assert any(record.levelno == logging.WARNING for record in caplog.records)
-
-    def test_a_signed_message_is_left_alone(self, tmp_path: Path):
-        """Replacing the body of a signed message would leave it claiming a
-        signature it no longer satisfies, and no extra request is worth making."""
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, message = self._client_writing(SIGNED_MESSAGE, "Only the newest sentence.")
-        download_path = tmp_path / "msg-1.eml"
-
-        with patch.object(OutlookConnectionConfig, "get_client", return_value=client):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert download_path.read_bytes() == SIGNED_MESSAGE
-        message.select.assert_not_called()
-
-    def test_no_staging_file_is_left_behind(self, tmp_path: Path):
-        """The reduced message is staged beside the target and moved into place,
-        so a later run must not find a stray partial file to reuse."""
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, _ = self._client_writing(raw, UNIQUE_HTML)
-        download_path = tmp_path / "msg-1.eml"
-
-        with patch.object(OutlookConnectionConfig, "get_client", return_value=client):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert [path.name for path in tmp_path.iterdir()] == ["msg-1.eml"]
-
-    @pytest.mark.parametrize("failing_call", ["write_bytes", "replace"], ids=["write", "replace"])
-    def test_an_unwritable_file_keeps_the_download_instead_of_failing(
-        self, tmp_path: Path, failing_call: str
-    ):
-        """The download on disk is already complete, so a disk failure must not
-        fail the record: one unwritable file would otherwise fail the whole run,
-        because a non-empty pipeline status raises at the end of it."""
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, _ = self._client_writing(raw, UNIQUE_HTML)
-        download_path = tmp_path / "msg-1.eml"
-
-        def boom(self, *args, **kwargs):
-            raise OSError("no space left on device")
-
-        with (
-            patch.object(OutlookConnectionConfig, "get_client", return_value=client),
-            patch.object(Path, failing_call, boom),
-        ):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert download_path.read_bytes() == raw
-        assert [path.name for path in tmp_path.iterdir()] == ["msg-1.eml"]
-
-    def test_cleanup_failure_after_a_staging_failure_does_not_fail_the_record(self, tmp_path: Path):
-        raw = HTML_AND_PLAIN_WITH_IMAGE_ATTACHMENT.read_bytes()
-        downloader = self._downloader(exclude_quoted_history=True)
-        client, _ = self._client_writing(raw, UNIQUE_HTML)
-        download_path = tmp_path / "msg-1.eml"
-
-        def fail_write(self, *args, **kwargs):
-            raise OSError("no space left on device")
-
-        def fail_cleanup(self, *args, **kwargs):
-            raise PermissionError("cannot remove staging file")
-
-        with (
-            patch.object(OutlookConnectionConfig, "get_client", return_value=client),
-            patch.object(Path, "write_bytes", fail_write),
-            patch.object(Path, "unlink", fail_cleanup),
-        ):
-            downloader._download_message(self._file_data(), download_path)
-
-        assert download_path.read_bytes() == raw
-
-
-class TestReduceMessageBodyKeepsBodyPartHeaders:
-    """set_content clears every Content-* header on the part it rewrites.
-
-    Those headers say what the part is and where it sits, not what it holds:
-    a multipart/related start= points at Content-ID, and Outlook mail resolves
-    relative references to inline resources through Content-Location.
-    """
-
-    RELATED_BODY = (
-        b"From: sender@example.com\r\nSubject: related\r\nMIME-Version: 1.0\r\n"
-        b'Content-Type: multipart/related; boundary="rel"; start="<body@x>"\r\n\r\n'
-        b"--rel\r\nContent-Type: text/html; charset=utf-8\r\n"
-        b"Content-ID: <body@x>\r\n"
-        b"Content-Location: http://example.invalid/mail.htm\r\n"
-        b"Content-Language: en-GB\r\n"
-        b"Content-Disposition: inline\r\n\r\n"
-        b"<html><body>quoted history</body></html>\r\n\r\n"
-        b"--rel\r\nContent-Type: image/png\r\nContent-ID: <img@x>\r\n"
-        b"Content-Transfer-Encoding: base64\r\n\r\naGk=\r\n\r\n--rel--\r\n"
-    )
-
-    @pytest.mark.parametrize(
-        "header",
-        ["Content-ID", "Content-Location", "Content-Language", "Content-Disposition"],
-    )
-    def test_the_header_survives_the_replacement(self, header: str):
-        reduced = _reduce_message_body(self.RELATED_BODY, UNIQUE_HTML)
-
-        body = _parse(reduced).get_body(preferencelist=BODY_PART_PREFERENCE)
-        assert (
-            body[header]
-            == _parse(self.RELATED_BODY).get_body(preferencelist=BODY_PART_PREFERENCE)[header]
-        )
-
-    def test_the_container_start_parameter_still_resolves(self):
-        reduced = _parse(_reduce_message_body(self.RELATED_BODY, UNIQUE_HTML))
-
-        start = reduced.get_param("start")
-        assert start is not None
-        assert any(part["Content-ID"] == start for part in reduced.iter_parts())
-
-    def test_no_mime_version_is_added_to_a_sub_part(self):
-        """RFC 2045 defines MIME-Version for the outermost entity only."""
-        reduced = _reduce_message_body(self.RELATED_BODY, UNIQUE_HTML)
-
-        assert reduced.count(b"MIME-Version") == 1
-
-
-class TestReduceMessageBodyWithTwoTextBlocks:
-    def test_a_second_narrative_text_part_is_dropped(self):
-        """Some clients write text, an image, then more text, side by side.
-
-        The second block is swept as though it were another rendering. Graph
-        returns the unique body as one piece covering both blocks and it is
-        written into the first part, so the text is moved rather than lost.
-        Pinned because nothing else states it, and because the sweep is where
-        a future change would silently start losing the text for real.
-        """
-        raw = (
-            b"From: sender@example.com\r\nSubject: two blocks\r\nMIME-Version: 1.0\r\n"
-            b'Content-Type: multipart/mixed; boundary="mix"\r\n\r\n'
-            b"--mix\r\nContent-Type: text/plain; charset=utf-8\r\n\r\nFIRST BLOCK\r\n\r\n"
-            b"--mix\r\nContent-Type: image/png\r\nContent-Transfer-Encoding: base64\r\n\r\n"
-            b"aGk=\r\n\r\n"
-            b"--mix\r\nContent-Type: text/plain; charset=utf-8\r\n\r\nSECOND BLOCK\r\n\r\n"
-            b"--mix--\r\n"
-        )
-
-        reduced = _reduce_message_body(raw, "the whole unique body")
-
-        assert b"SECOND BLOCK" not in reduced
-        assert b"the whole unique body" in reduced
-        assert b"image/png" in reduced
-
-
-class TestReduceMessageBodyLeavesProtectedMessagesAlone:
-    def test_a_signed_message_declines(self):
-        assert _reduce_message_body(SIGNED_MESSAGE, UNIQUE_HTML) is None
-
-    def test_an_encrypted_message_declines(self):
-        encrypted = SIGNED_MESSAGE.replace(b"multipart/signed", b"multipart/encrypted")
-
-        assert _reduce_message_body(encrypted, UNIQUE_HTML) is None
